chore(search-api): Move DatasetsSearchController to package no.fdk.searchapi.controller.datasetssearch in order to prepare split refactoring.
