+++ applications/search-api/src/main/java/no/fdk/searchapi/controller/datasetssearch/DatasetsSearchController.java
@@ -1,4 +1,4 @@
-package no.fdk.searchapi.controller;
+package no.fdk.searchapi.controller.datasetssearch;
 
 import com.google.common.collect.ImmutableMap;
 import io.swagger.annotations.ApiImplicitParam;
